Fixed feeding recognition bug
