--- Assets/Prefabs/Foundation_Ramp Variant.prefab
@@ -1,86 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!1001 &5064149085849883958
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalScale.x
-      value: 12
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalScale.y
-      value: 12
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalScale.z
-      value: 12
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: -2.898543
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: -0.00499988
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: -5.430418
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 919132149155446097, guid: feda470e680b1ea4ba1c0c05f3d16a8c,
-        type: 3}
-      propertyPath: m_Name
-      value: Foundation_Ramp
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: feda470e680b1ea4ba1c0c05f3d16a8c, type: 3}
