+++ Assets/Prefabs/LP_SmallCreature Variant.prefab
@@ -84,3 +84,21 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: e5332f561130ddd45b51fe8fc9736245, type: 3}
+--- !u!1 &226812425662747705 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: e5332f561130ddd45b51fe8fc9736245,
+    type: 3}
+  m_PrefabInstance: {fileID: 1145222049514973544}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &311103127921467463
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 226812425662747705}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f164746d1950a5c49a9ecfefe99ad6cf, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
