+++ Assets/Prefabs/MobileMidCreature.prefab
@@ -221,3 +221,15 @@ MonoBehaviour:
     serializedVersion: 2
     m_Bits: 4294967291
   slopeLimit: 80
+--- !u!114 &5089405816611752533
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2223632638037285957}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f164746d1950a5c49a9ecfefe99ad6cf, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
