+++ Assets/Prefabs/UpdatedCreatureBase.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c896534d2b039824c9aa44fbe035ec2c
+guid: d32f67bbc6420d04390d8d755413ac89
 PrefabImporter:
   externalObjects: {}
   userData:
