+++ Assets/Scripts/Creature/CollisionRelay.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f164746d1950a5c49a9ecfefe99ad6cf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
