+++ UserSettings/EditorUserSettings.asset
@@ -15,22 +15,22 @@ EditorUserSettings:
       value: 224247031146466e0307183b3d2c49151b12163e0d01691535281023eee52777b1c379dee6343b392b4cfa320d2a18
       flags: 0
     RecentlyUsedScenePath-3:
-      value: 22424703114646680e0b0227036c6f1c17141d052e222333393a3c01acf53a31f6fe
-      flags: 0
-    RecentlyUsedScenePath-4:
       value: 224247031146466e0307183b3d2c49151b12163e0d01691535281023eee52777b1c379cfe83633332e32ee280c701431fb1e10
       flags: 0
-    RecentlyUsedScenePath-5:
+    RecentlyUsedScenePath-4:
       value: 224247031146466e0307183b3d2c49151b12163e0d01691535281023eee52777b1c379dee636331d2f0de6380530023aa0441c05ff1f13
       flags: 0
-    RecentlyUsedScenePath-6:
+    RecentlyUsedScenePath-5:
       value: 224247031146466e0307183b3d2c49151b12163e0d01691535281023eee52777b1c379dee636331d2f0de6380530023aa3441c05ff1f13
       flags: 0
-    RecentlyUsedScenePath-7:
+    RecentlyUsedScenePath-6:
       value: 22424703114646680e0b0227036c6c1817131d38182d35241e2a183de7ae2136ebf32f
       flags: 0
-    RecentlyUsedScenePath-8:
+    RecentlyUsedScenePath-7:
       value: 22424703114646680e0b0227036c7c0213160c3f3e2d0b3f3b2c1036ecf40434e3fe31fbe82f31387717e1351027
+      flags: 0
+    RecentlyUsedScenePath-8:
+      value: 22424703114646680e0b0227036c6f1c17141d052e222333393a3c01acf53a31f6fe
       flags: 0
     RecentlyUsedScenePath-9:
       value: 22424703114646680e0b0227036c7a06191b0d3e2527281d222a1600e1e53a3dacf238e0f323
